backend/requirements.txt
faiss-cpu
openai
sentence-transformers
PyPDF2
python-dotenv
uvicorn
fastapi
python-multipart
gtts
whisper
openai-whisper 
imageio[ffmpeg]
google-generativeai
cohere
SpeechRecognition